+++ src/components/SignUpForm/SignUpForm.js
@@ -2,8 +2,9 @@ import React, { Component } from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import "./signup.css"
 
-import { withFirebase } from '../../pages/Firebase'
+import { withFirebase } from '../Firebase'
 import * as ROUTES from '../../constants/routes'
+
 
 const INIT_STATE = {
   username: '',
