+++ src/pages/SignUp/index.js
@@ -1,10 +1,6 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import React from 'react';
 import SignUpForm  from '../../components/SignUpForm/SignUpForm'
 
-import { withFirebase } from '../Firebase'
-
-import * as ROUTES from '../../constants/routes'
 
 const SignUpPage = () => (
   <div>
